Initial changes for alarm clock implementation

+++ src/lib/datastructures.c
@@ -2,7 +2,7 @@
 #include "datastructures.h"
 #include <stdlib.h>
 
-Heap heap_init(int32_t (*key) (void*)) {
+Heap heap_init(int64_t (*key) (void*)) {
     Heap heap = (Heap) malloc (sizeof(struct heap));
     heap->elems = (void**) malloc (HEAP_CAPACITY_DEFAULT * sizeof(void*));
     heap->capacity = HEAP_CAPACITY_DEFAULT;
@@ -14,7 +14,7 @@ Heap heap_init(int32_t (*key) (void*)) {
 
 void heap_double_capacity(Heap heap) {  
     void** elems_copy = (void**) malloc (heap->capacity * 2 * sizeof(void*));
-    uint32_t i;
+    uint64_t i;
     for (i = 1; i <= heap->size; i++) {
         elems_copy[i] = heap->elems[i];
     }
@@ -35,11 +35,11 @@ void heap_insert_elem(Heap heap, void* elem) {
     heap->size++;
 
     if (heap->size > 1) {
-        uint32_t child_idx = heap->size;
-        uint32_t parent_idx = child_idx / 2;
+        uint64_t child_idx = heap->size;
+        uint64_t parent_idx = child_idx / 2;
         while(parent_idx > 0) {
-            int32_t parent_key = heap->key(heap->elems[parent_idx]);
-            int32_t child_key = heap->key(heap->elems[child_idx]);
+            int64_t parent_key = heap->key(heap->elems[parent_idx]);
+            int64_t child_key = heap->key(heap->elems[child_idx]);
             if (child_key >= parent_key) {
                 return;
             }
@@ -72,18 +72,18 @@ void* heap_pop_min(Heap heap) {
         elems[heap->size] = NULL;
         heap->size--;
 
-        uint32_t parent_idx = 1;
-        uint32_t left_idx = parent_idx * 2;
+        uint64_t parent_idx = 1;
+        uint64_t left_idx = parent_idx * 2;
 
         while (left_idx <= heap->size) {
-            int32_t parent_key = heap->key(elems[parent_idx]);
-            int32_t left_key = heap->key(elems[left_idx]);
-            uint32_t min_idx = (left_key < parent_key) ? left_idx : parent_idx;
-            int32_t min_key = (left_key < parent_key) ? left_key : parent_key;
-            uint32_t right_idx = left_idx + 1;
+            int64_t parent_key = heap->key(elems[parent_idx]);
+            int64_t left_key = heap->key(elems[left_idx]);
+            uint64_t min_idx = (left_key < parent_key) ? left_idx : parent_idx;
+            int64_t min_key = (left_key < parent_key) ? left_key : parent_key;
+            uint64_t right_idx = left_idx + 1;
 
             if (right_idx <= heap->size) {
-                int32_t right_key = heap->key(elems[right_idx]);
+                int64_t right_key = heap->key(elems[right_idx]);
                 min_idx = (right_key < min_key) ? right_idx : min_idx;
             }
 
@@ -100,7 +100,7 @@ void* heap_pop_min(Heap heap) {
     return min_elem;
 }
 
-void heap_swap_elems(Heap heap, uint32_t idx1, uint32_t idx2) {
+void heap_swap_elems(Heap heap, uint64_t idx1, uint64_t idx2) {
     if ((idx1 <= heap->size) && (idx2 <= heap->size)) {
         void* tmp = heap->elems[idx1];
         heap->elems[idx1] = heap->elems[idx2];
